requirements.txt
absl-py==0.15.0
astunparse==1.6.3
beautifulsoup4==4.10.0
boto3==1.18.33
botocore==1.21.33
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
deepface==0.0.68
filelock==3.3.2
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mtcnn==0.1.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.2
Pillow==8.3.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032501
threadpoolctl==2.2.0
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
